+++ example/lib/decoration/barrel_dragable.dart
@@ -16,14 +16,23 @@ class BarrelDraggable extends GameDecoration with DragGesture, ObjectCollision {
     setupCollision(
       CollisionConfig(
         collisions: [
-          CollisionArea(
-            width: DungeonMap.tileSize * 0.6,
-            height: DungeonMap.tileSize * 0.4,
-            align: Offset(
+          CollisionArea.rectangle(
+            size: Size(
+              DungeonMap.tileSize * 0.6,
+              DungeonMap.tileSize * 0.4,
+            ),
+            align: Vector2(
               DungeonMap.tileSize * 0.2,
               DungeonMap.tileSize * 0.4,
             ),
-          )
+          ),
+          // CollisionArea.circle(
+          //   radius: (DungeonMap.tileSize) / 2,
+          // align: Vector2(
+          //   DungeonMap.tileSize * 0.2,
+          //   DungeonMap.tileSize * 0.4,
+          // ),
+          // ),
         ],
       ),
     );