+++ example/lib/decoration/column.dart
@@ -1,7 +1,8 @@
+import 'dart:ui';
+
 import 'package:bonfire/bonfire.dart';
 import 'package:example/manual_map/dungeon_map.dart';
 import 'package:example/util/common_sprite_sheet.dart';
-import 'package:flutter/cupertino.dart';
 
 class ColumnDecoration extends GameDecoration with ObjectCollision {
   ColumnDecoration(Vector2 position)
@@ -14,10 +15,9 @@ class ColumnDecoration extends GameDecoration with ObjectCollision {
     setupCollision(
       CollisionConfig(
         collisions: [
-          CollisionArea(
-            width: DungeonMap.tileSize,
-            height: DungeonMap.tileSize / 2,
-            align: Offset(0, DungeonMap.tileSize * 1.8),
+          CollisionArea.rectangle(
+            size: Size(DungeonMap.tileSize, DungeonMap.tileSize / 2),
+            align: Vector2(0, DungeonMap.tileSize * 1.8),
           ),
         ],
       ),