+++ lib/map/map_game.dart
@@ -1,6 +1,7 @@
 import 'dart:ui';
 
 import 'package:bonfire/base/bonfire_game.dart';
+import 'package:bonfire/collision/object_collision.dart';
 import 'package:bonfire/map/tile/tile.dart';
 import 'package:bonfire/util/priority_layer.dart';
 import 'package:flame/components.dart';
@@ -14,8 +15,8 @@ abstract class MapGame extends Component with HasGameRef<BonfireGame> {
 
   Iterable<Tile> getRendered();
 
-  Iterable<Tile> getCollisionsRendered();
-  Iterable<Tile> getCollisions();
+  Iterable<ObjectCollision> getCollisionsRendered();
+  Iterable<ObjectCollision> getCollisions();
 
   Future<void> updateTiles(Iterable<Tile> map);