+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-# NEXT
+# 1.0.0-rc4
 
-- fix to use `Tap Gesture` in `Drag Gesture` using together.
+- BREAKING CHANGE: add Shapes(circle,rectangle,polygon) to use collisions.
+- fix to use `TapGesture` in `DragGesture` using together.
+- Improvements in mixin `Sensor`.
+- Improvements in `TalkDialog`.
 
 # 1.0.0-rc3